# Performance optimizations for querying for educator authorizations