--- a/app/controllers/admin/educators_controller.rb
+++ b/app/controllers/admin/educators_controller.rb
@@ -21,12 +21,12 @@ module Admin
 
     def authorization
       @all_educators = Educator.all
-      
+
       @districtwide_educators = []
       @can_set_educators = []
       @admin_educators = []
       @restricted_notes_educators = []
-    
+
       @all_educators.each do |educator|
         @districtwide_educators << educator if educator.districtwide_access
         @can_set_educators << educator if educator.can_set_districtwide_access